Conditional compilation + test rabbit hole
Move conditionally-compiled symbols from unit test into own file.

Also disable recursive dev-dependency, which causes following dev-dependency cycle during cargo test.
Note that #[cfg(test)] is not carried along dependencies.

  godot-core [dev]
  -> godot [non-dev]
  -> godot-core [non-dev]

So, we get a Heisenberg situation where both #[cfg(test)] and #[cfg(not(test))] are simultaneously true.
This also happens without any dependencies, just thanks to doctest enabling #[cfg(doctest)] but not #[cfg(test)].

The following code prints *both* errors!

  #[cfg(test)]
  compile_error!("test");

  #[cfg(not(test))]
  compile_error!("not(test)");

See comments in godot-core/lib.rs for more horror stories.
