+++ godot-core/src/lib.rs
@@ -4,22 +4,38 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+// If running in tests, a lot of symbols are unused or panic early
+#![cfg_attr(feature = "unit-test", allow(unreachable_code, unused))]
+
+// More test hacks...
+//
+// Technically, `cargo test -p godot-core` *could* be supported by this abomination:
+//   #[cfg(not(any(test, doctest, feature = "unit-test"))]
+// which would be necessary because `cargo test` runs both test/doctest, and downstream crates may need the feature as
+// workaround https://github.com/rust-lang/rust/issues/59168#issuecomment-962214945. However, this *also* does not work,
+// as #[cfg(doctest)] is currently near-useless for conditional compilation: https://github.com/rust-lang/rust/issues/67295.
+// As if this weren't enough, even our compile error here does not appear as the first error message, but the 4th or so.
+#[cfg(all(test, not(feature = "unit-test")))]
+compile_error!("Running `cargo test` requires `--features unit-test`.");
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
 mod registry;
 mod storage;
 
 pub mod bind;
 pub mod builder;
 pub mod builtin;
-#[cfg(not(feature = "unit-test"))]
-pub mod engine;
 pub mod init;
 pub mod log;
 pub mod macros;
 pub mod obj;
 
+pub use godot_ffi as sys;
 pub use registry::*;
 
-pub use godot_ffi as sys;
+#[cfg(not(feature = "unit-test"))]
+pub mod engine;
 
 // Output of generated code. Mimics the file structure, symbols are re-exported.
 #[cfg(not(feature = "unit-test"))]
@@ -30,119 +46,10 @@ mod gen {
     pub mod utilities;
 }
 
-#[cfg(feature = "unit-test")]
-mod gen {
-    pub mod central {
-        pub mod global {}
-    }
-    pub mod classes {
-        pub struct Node {}
-        pub struct Resource {}
-
-        pub mod class_macros {}
-    }
-    pub mod utilities {}
-}
-
-#[cfg(feature = "unit-test")]
-pub mod engine {
-    use super::sys;
-    use crate::obj::{Gd, GodotClass};
-
-    pub struct Object {}
-    pub struct RefCounted {}
-    impl RefCounted {
-        pub fn init_ref(&self) -> bool {
-            sys::panic_no_godot!(RefCounted::init_ref)
-        }
-        pub fn reference(&self) -> bool {
-            sys::panic_no_godot!(RefCounted::reference)
-        }
-        pub fn unreference(&self) -> bool {
-            sys::panic_no_godot!(RefCounted::unreference)
-        }
-    }
-
-    impl GodotClass for Object {
-        type Base = ();
-        type Declarer = crate::obj::dom::EngineDomain;
-        type Mem = crate::obj::mem::DynamicRefCount;
-        const CLASS_NAME: &'static str = "";
-    }
-    impl GodotClass for RefCounted {
-        type Base = Object;
-        type Declarer = crate::obj::dom::EngineDomain;
-        type Mem = crate::obj::mem::StaticRefCount;
-        const CLASS_NAME: &'static str = "";
-    }
-
-    pub mod utilities {
-        use super::sys;
-
-        pub fn is_instance_id_valid(id: i64) -> bool {
-            sys::panic_no_godot!(is_instance_id_valid)
-        }
-    }
-
-    #[allow(non_camel_case_types)]
-    pub mod global {
-        use super::sys;
-
-        #[derive(Debug)]
-        pub enum PropertyHint {
-            PROPERTY_HINT_NONE,
-        }
-        impl PropertyHint {
-            pub fn ord(&self) -> i32 {
-                sys::panic_no_godot!(PropertyHint::ord)
-            }
-        }
-
-        #[derive(Debug)]
-        pub enum PropertyUsageFlags {
-            PROPERTY_USAGE_DEFAULT,
-        }
-        impl PropertyUsageFlags {
-            pub fn ord(&self) -> i32 {
-                sys::panic_no_godot!(PropertyUsageFlags::ord)
-            }
-        }
-    }
-
-    pub(crate) fn debug_string<T: GodotClass>(
-        ptr: &Gd<T>,
-        f: &mut std::fmt::Formatter<'_>,
-        ty: &str,
-    ) -> std::fmt::Result {
-        sys::panic_no_godot!(Debug)
-    }
-
-    pub(crate) fn display_string<T: GodotClass>(
-        ptr: &Gd<T>,
-        f: &mut std::fmt::Formatter<'_>,
-    ) -> std::fmt::Result {
-        sys::panic_no_godot!(Display)
-    }
-}
-
-pub mod callbacks {
-    use super::sys;
-    use crate::obj::{Base, GodotClass};
-
-    pub unsafe extern "C" fn create<T>(
-        _class_userdata: *mut std::ffi::c_void,
-    ) -> sys::GDNativeObjectPtr {
-        sys::panic_no_godot!(create)
-    }
-
-    pub(crate) fn create_custom<T, F>(_make_user_instance: F) -> sys::GDNativeObjectPtr
-    where
-        T: GodotClass,
-        F: FnOnce(Base<T::Base>) -> T,
-    {
-        sys::panic_no_godot!(create_custom)
-    }
-}
+#[cfg(any(feature = "unit-test"))]
+mod test_stubs;
+#[cfg(any(feature = "unit-test"))]
+use test_stubs::*;
 
 #[doc(hidden)]
 pub mod private {
